Revert "FVR: Fixes to curvy vs pointy"
This reverts commit dd4370f2408013a00655cd202e846add5bed93fd.

+++ Assets/Scripts/Free Game/FreeGlobals.cs
@@ -84,7 +84,7 @@ public static class FreeGlobals
 	public static float choiceDelay = 0;
 	public static float startRewardDelay = 0;
 	public static string rewardedOri = "none";
-	public static int firstRewardDur = -1; // ms
+	public static float firstRewardDur = -1; // ms
 
 	// Parameters for curvy vs. straight discrimination
 	public static string rewardedLineType;
